backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
opencv-python-headless
numpy
python-multipart
